[CHE-781] Bump. Move code from extension view controller within the view controller (no need to split a less than 200 lines file). Move not-tableview related functions outside of the table view delegate & datasource extension. Add missing super.viewWillDisappear(animated) on overriding. Add fixme comment for later.
